HDDS-11728. Refactor subcommand layouts of ozone debug and repair
