+++ hadoop-ozone/integration-test/src/test/java/org/apache/hadoop/ozone/parser/TestOzoneHARatisLogParser.java
@@ -29,8 +29,8 @@ import org.apache.hadoop.ozone.client.OzoneClient;
 import org.apache.hadoop.ozone.client.OzoneClientFactory;
 import org.apache.hadoop.ozone.om.helpers.OMRatisHelper;
 import org.apache.hadoop.ozone.MiniOzoneCluster;
-import org.apache.hadoop.ozone.segmentparser.OMRatisLogParser;
-import org.apache.hadoop.ozone.segmentparser.SCMRatisLogParser;
+import org.apache.hadoop.ozone.debug.segmentparser.OMRatisLogParser;
+import org.apache.hadoop.ozone.debug.segmentparser.SCMRatisLogParser;
 import org.apache.ozone.test.GenericTestUtils;
 import org.apache.ozone.test.tag.Flaky;
 import org.junit.jupiter.api.AfterEach;
