+++ hadoop-ozone/tools/src/main/java/org/apache/hadoop/ozone/debug/chunk/ChunkDataNodeDetails.java
@@ -1,4 +1,4 @@
-/**
+/*
  * Licensed to the Apache Software Foundation (ASF) under one
  * or more contributor license agreements.  See the NOTICE file
  * distributed with this work for additional information
@@ -6,17 +6,38 @@
  * to you under the Apache License, Version 2.0 (the
  * "License"); you may not use this file except in compliance
  * with the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
+ * <p>
+ * http://www.apache.org/licenses/LICENSE-2.0
+ * <p>
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.hadoop.ozone.segmentparser;
 
+package org.apache.hadoop.ozone.debug.chunk;
 /**
- * Command line utility for dump ratis log files.
+ * Class that gives datanode details on which the chunk is present.
  */
+public class ChunkDataNodeDetails {
+  private String ipAddress;
+  private String hostName;
+
+  public ChunkDataNodeDetails(String ipAddress, String hostName) {
+    this.ipAddress = ipAddress;
+    this.hostName = hostName;
+  }
+
+  @Override
+    public String toString() {
+    return "{"
+            + "ipAddress='"
+            + ipAddress
+            + '\''
+            + ", hostName='"
+            + hostName
+            + '\''
+            + '}';
+  }
+}
