+++ hadoop-ozone/tools/src/main/java/org/apache/hadoop/ozone/debug/chunk/ChunkDetails.java
@@ -6,9 +6,9 @@
  * to you under the Apache License, Version 2.0 (the
  * "License"); you may not use this file except in compliance
  * with the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
+ * <p>
+ * http://www.apache.org/licenses/LICENSE-2.0
+ * <p>
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -16,30 +16,39 @@
  * limitations under the License.
  */
 
-package org.apache.hadoop.ozone.segmentparser;
-
-import org.apache.hadoop.hdds.cli.HddsVersionProvider;
-import org.apache.hadoop.ozone.om.helpers.OMRatisHelper;
-import picocli.CommandLine;
-
-import java.util.concurrent.Callable;
+package org.apache.hadoop.ozone.debug.chunk;
 
 /**
- * Command line utility to parse and dump a OM ratis segment file.
+ * Class that gives chunkDetails.
  */
-@CommandLine.Command(
-    name = "om",
-    description = "dump om ratis segment file",
-    mixinStandardHelpOptions = true,
-    versionProvider = HddsVersionProvider.class)
-public class OMRatisLogParser extends BaseLogParser implements Callable<Void> {
+public class ChunkDetails {
+  private String chunkName;
+  private long chunkOffset;
+
+  public String getChunkName() {
+    return chunkName;
+  }
+
+  public void setChunkName(String chunkName) {
+    this.chunkName = chunkName;
+  }
 
   @Override
-  public Void call() throws Exception {
-    System.out.println("Dumping OM Ratis Log");
+    public String toString() {
+    return "{"
+            + "chunkName='"
+            + chunkName
+            + '\''
+            + ", chunkOffset="
+            + chunkOffset
+            + '}';
+  }
 
-    parseRatisLogs(OMRatisHelper::smProtoToString);
-    return null;
+  public long getChunkOffset() {
+    return chunkOffset;
+  }
+
+  public void setChunkOffset(long chunkOffset) {
+    this.chunkOffset = chunkOffset;
   }
 }
-
