+++ hadoop-ozone/tools/src/main/java/org/apache/hadoop/ozone/debug/segmentparser/BaseLogParser.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.hadoop.ozone.segmentparser;
+package org.apache.hadoop.ozone.debug.segmentparser;
 
 import com.google.common.annotations.VisibleForTesting;
 import org.apache.ratis.proto.RaftProtos;
